+++ CHANGELOG.md
@@ -5,3 +5,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 - Initial release
+
+## [0.1.0] - 2017-08-05
+### Added
+- Add macro include guard to cpp header files
+- Generate getters and setters